# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSync API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_api_v1_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/medications": {
      "get": {
        "tags": [
          "medication"
        ],
        "summary": "List Medications",
        "operationId": "list_medications_api_v1_medications_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedicamentoDto"
                  },
                  "title": "Response List Medications Api V1 Medications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "medication"
        ],
        "summary": "Add Medication",
        "operationId": "add_medication_api_v1_medications_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedicationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicamentoCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/medications/{id_medicamento}": {
      "put": {
        "tags": [
          "medication"
        ],
        "summary": "Update Medication",
        "operationId": "update_medication_api_v1_medications__id_medicamento__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_medicamento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Medicamento"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMedicationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicamentoCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "medication"
        ],
        "summary": "Remove Medication",
        "operationId": "remove_medication_api_v1_medications__id_medicamento__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_medicamento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Medicamento"
            }
          },
          {
            "name": "fecha_fin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Fin"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intakes": {
      "get": {
        "tags": [
          "medication"
        ],
        "summary": "Get Intakes",
        "operationId": "get_intakes_api_v1_intakes_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TomaDto"
                  },
                  "title": "Response Get Intakes Api V1 Intakes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intakes/{id_toma}/confirm": {
      "post": {
        "tags": [
          "medication"
        ],
        "summary": "Confirm Intake",
        "operationId": "confirm_intake_api_v1_intakes__id_toma__confirm_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_toma",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Toma"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TomaDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intakes/{id_toma}/postpone": {
      "post": {
        "tags": [
          "medication"
        ],
        "summary": "Postpone Intake",
        "operationId": "postpone_intake_api_v1_intakes__id_toma__postpone_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_toma",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Toma"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TomaDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vital-signs/latest": {
      "get": {
        "tags": [
          "vital-signs"
        ],
        "summary": "Get Latest",
        "operationId": "get_latest_api_v1_vital_signs_latest_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignoVitalDto"
                  },
                  "title": "Response Get Latest Api V1 Vital Signs Latest Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vital-signs/history": {
      "get": {
        "tags": [
          "vital-signs"
        ],
        "summary": "Get History",
        "operationId": "get_history_api_v1_vital_signs_history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To"
            }
          },
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignoVitalDto"
                  },
                  "title": "Response Get History Api V1 Vital Signs History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vital-signs": {
      "post": {
        "tags": [
          "vital-signs"
        ],
        "summary": "Create Reading",
        "operationId": "create_reading_api_v1_vital_signs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVitalSignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SignoVitalDto"
                  },
                  "type": "array",
                  "title": "Response Create Reading Api V1 Vital Signs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMedicationRequest": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "dosis": {
            "type": "string",
            "title": "Dosis"
          },
          "frecuencia_horas": {
            "type": "integer",
            "title": "Frecuencia Horas"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Fin"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "dosis",
          "frecuencia_horas",
          "fecha_inicio"
        ],
        "title": "CreateMedicationRequest"
      },
      "CreateVitalSignRequest": {
        "properties": {
          "frecuencia_cardiaca": {
            "type": "integer",
            "maximum": 300.0,
            "minimum": 0.0,
            "title": "Frecuencia Cardiaca"
          },
          "spo2": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Spo2"
          },
          "calidad_senal": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Calidad Senal"
          },
          "registrado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Registrado En"
          }
        },
        "type": "object",
        "required": [
          "frecuencia_cardiaca",
          "spo2",
          "calidad_senal",
          "registrado_en"
        ],
        "title": "CreateVitalSignRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedicamentoCreateResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "dosis": {
            "type": "string",
            "title": "Dosis"
          },
          "frecuencia_horas": {
            "type": "integer",
            "title": "Frecuencia Horas"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Fin"
          },
          "advertencia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advertencia"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "dosis",
          "frecuencia_horas",
          "fecha_inicio"
        ],
        "title": "MedicamentoCreateResponse",
        "description": "Mismas claves que MedicamentoDto más una `advertencia` opcional de\nsuperposición horaria (HU-006, criterio 4); se omite si no aplica\n(response_model_exclude_none=True en la ruta) para no romper el\n`fromJson` del frontend, que ignora claves desconocidas."
      },
      "MedicamentoDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "dosis": {
            "type": "string",
            "title": "Dosis"
          },
          "frecuencia_horas": {
            "type": "integer",
            "title": "Frecuencia Horas"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Fin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "dosis",
          "frecuencia_horas",
          "fecha_inicio"
        ],
        "title": "MedicamentoDto"
      },
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "tipo_perfil": {
            "type": "string",
            "title": "Tipo Perfil"
          },
          "fecha_nacimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Nacimiento"
          },
          "patologias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patologias"
          },
          "parentesco": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentesco"
          },
          "matricula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matricula"
          },
          "especialidad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Especialidad"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "nombre",
          "apellido",
          "tipo_perfil"
        ],
        "title": "ProfileResponse"
      },
...
      "SignoVitalDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "valor": {
            "type": "number",
            "title": "Valor"
          },
          "fecha_medicion": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Medicion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tipo",
          "valor",
          "fecha_medicion"
        ],
        "title": "SignoVitalDto"
      },
      "TomaDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fecha_programada": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Programada"
          },
          "fecha_confirmada": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Confirmada"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "horario_id": {
            "type": "integer",
            "title": "Horario Id"
          },
          "nombre_medicamento": {
            "type": "string",
            "title": "Nombre Medicamento"
          },
          "dosis": {
            "type": "string",
            "title": "Dosis"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fecha_programada",
          "estado",
          "horario_id",
          "nombre_medicamento",
          "dosis"
        ],
        "title": "TomaDto"
      },
      "UpdateMedicationRequest": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "dosis": {
            "type": "string",
            "title": "Dosis"
          },
          "frecuencia_horas": {
            "type": "integer",
            "title": "Frecuencia Horas"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Fin"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "dosis",
          "frecuencia_horas",
          "fecha_inicio"
        ],
        "title": "UpdateMedicationRequest"
      },
      "UpdateProfileRequest": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "fecha_nacimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Nacimiento"
          },
          "patologias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patologias"
          },
          "parentesco": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentesco"
          },
          "matricula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matricula"
          },
          "especialidad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Especialidad"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "apellido"
        ],
        "title": "UpdateProfileRequest"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        }
      }
    }
  }
}